parse_future_committee_meetings: use Israel timezone
icalendar files are by default in UTC. This causes the current scraped
files, that were implicitly in Isr timezone, to be 2 hours in the
future. There are two ways to fix this:
 1. create the icalendar files assuming the db is in Israel timezone.
 2. change the db to be in UTC timezone.

To allow easier cooperation with external data the later approach is
taken. This patch fixes the scraping code for future committees. Similar
patches need to be applied to the whole tree, for any model instance
creation or change of a datetime or time field. Additionally data
migration will be written to change an existing database and move all
times back, based on the date (since some will be in daylight savings
and some not).